backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.4
httpx==0.26.0
psycopg2-binary==2.9.9

# Optional: For AI features
# openai==1.10.0
# langchain==0.1.0

# Groq AI (Free and Fast)
groq==0.4.1
